requirements.txt
langchain==0.1.0
langgraph==0.0.20
langchain-openai==0.0.5
langchain-community==0.0.10
openai==1.3.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
python-dotenv==1.0.0
pydantic==2.4.2
typing-extensions==4.7.1
# Evaluation dependencies
nltk==3.8.1
matplotlib==3.7.2
seaborn==0.12.2
